Created Variable to use for both SAST and SCA
Added a variable called $(Application_Profile_name) that stores the app profile name as stated on Veracode and referenced it in both the SAST and SCA steps to ensure that the name matches.

Also added "import flaws" to view flaws in the pipeline after the SAST scan, and updated the version of jackson-databind to a vulnerable version to demonstrate vetting SCA scans.